AISDK-191: Add topic extraction client
