+++ src/test/java/ai/rev/speechtotext/unit/RevAiTranscriptTest.java
@@ -1,8 +1,8 @@
 package ai.rev.speechtotext.unit;
 
+import ai.rev.helpers.MockInterceptor;
 import ai.rev.speechtotext.ApiInterface;
 import ai.rev.speechtotext.ApiClient;
-import ai.rev.speechtotext.MockInterceptor;
 import ai.rev.speechtotext.models.asynchronous.Element;
 import ai.rev.speechtotext.models.asynchronous.Monologue;
 import ai.rev.speechtotext.models.asynchronous.RevAiTranscript;
